refactor(theme): create VuetifyThemeAdapter for CSS generation
Extends v0's ThemeAdapter to produce Vuetify-specific layered CSS
(@layer vuetify-utilities) with RGB-decomposed custom properties,
overlay multipliers, and bg/text/border utility classes. Handles
both browser DOM injection via adoptedStyleSheets and SSR via @unhead.
